+++ src/infrastructure/decorators/typeorm/fields/TypeOrmImageField/index.ts
@@ -3,7 +3,7 @@ import {IFileField} from '../../../fields/ImageField';
 
 export default (options: IFileField) => [
     Column({
-        type: options.multiple ? 'simple-array' : 'integer',
+        type: options.isArray ? 'simple-array' : 'integer',
         default: options.defaultValue,
         nullable: options.nullable,
     }),